Avoiding parallel error handling w/ update_grid()
It could totally stop the script at the end because it'll trigger both the AJAX error handling if update_grid() fails and the empty response handling in start_round() and end_round(). So AJAX error handling is disabled specifically in the calls in start_round() and end_round(). One error handling is already enough :p.