[Minor] Saving some seconds in the planets checks
When we're in a difficulty inferior to the one wanted (currently 3, highest known), a check for a better planet is executed after we succesfully started a new round. Previously if our planet is the current best one, it'd get a better zone and switch while it's not needed. So i added a timer check inside `CheckSwitchBetterPlanet()` to do nothing if it's a difficulty check and a round started <= 8s ago.
It'll more or less save 2/3s every 5 completed rounds.