Avoiding an useless planet switch
A rare case could happen where a zone (of a difficulty +1 vs all the others difficulties available) just completed. It'd check the planets available and since auto_switch_planet.current_difficulty would still be equal to 2, it'd not return the current planet as the best one because the max difficulty of others planets would be different.